requirements.txt
discord.py>=2.3.0
motor>=3.3.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
